+++ src/components/SprintSidebar/SprintSidebar.module.css
@@ -30,8 +30,13 @@
   padding: 0;
   border-radius: 50%;
   cursor: pointer;
+  background-color: #e5e5e5;
+  
 } 
 .button_plus {
+    margin-left: auto;
+    margin-right: auto;
+    margin-bottom: 15px;
     display: flex;
     justify-content: center;
     align-items: center;
