+++ src/components/SprintTaskList/SprintTaskList.module.scss
@@ -13,6 +13,10 @@
     background: #FFFFFF;
     box-shadow: 0px 6px 26px rgba(0, 5, 97, 0.1);
     border-radius: 5px;
+    transition: all 200ms cubic-bezier(0.175, 0.885, 0.32, 1.275);
+    &:hover {
+      transform: scale(1.1, 1.1);
+    }
 }
 .task_details {
     display: block;
@@ -21,10 +25,23 @@
 }
 .task_details_item {
     display: flex;
+    justify-content: space-between;
     align-items: center;
+    margin-bottom: 20px;
     border-bottom: 1px solid rgba(24, 28, 39, 0.2);
-
+    // &::after {
+    //     content: '';
+    //     display: flex;
+    //     position: absolute;
+    //     left: 0;
+    //     bottom: 0;
+    //     width: 100%;
+    //     height: 1px;
+    //     border-radius: 2px;
+    //     background-color: #A6A6A6;
+    // }
 }
+
 .task_details_item:last-child {
     border: none;
 }
@@ -32,26 +49,31 @@
 @media screen and (min-width: 768px) {
     .task_card {
         width: 216px;
-        /* height: 284px; */
-        /* padding: 20px 30px 5px; */
     }
-
     .task_details_item {
         display: flex;
         align-items: center;
-        /* border-bottom: 1px solid rgba(24, 28, 39, 0.2); */
-    
+    }
+    .task_list {
+        font-size: 18px;
+        line-height: 22px;
     }
 }
 @media screen and (min-width: 1200px) {
     .task_card {
         width: 869px;
-        height: 95px;
     }
-        .task_details {
+    .task_details {
         display: flex;
         justify-content: space-between;
         align-items: center;
+    }
+    .task_details_item {
+        width: 100px;
+        margin-bottom: 0px;
+    }
+    .hide_title {
+        display: none;
     }
 
 }
@@ -64,15 +86,20 @@
     font-family: "Montserrat";
     font-style: normal;
     font-weight: 500;
-    font-size: 18px;
-    line-height: 22px;
-    letter-spacing: .04em;
+    font-size: 14px;
+    line-height: 17px;
+    letter-spacing: 0.04em;
     color: #181c27;
     max-width: 100%;
+    margin: auto;
+}
+.task_details_item:first-child {
+    font-size: 18px;
+    line-height: 22px;
 }
 input {
-    width: 176px;
-    height: 58px;
+    width: 13px;
+    // height: 58px;
     border: none;
     outline: none;
 }
@@ -83,8 +110,8 @@ input {
     position: absolute;
     bottom: 10px;
     right: 10px;
-    width: 30px;
-    height: 30px;
+    width: 20px;
+    height: 20px;
     background: #A6A6A6 ;
     border: none;
     border-radius: 50%;
@@ -95,8 +122,8 @@ input {
 
 }
 .button_delete {
-    width: 20px;
-    height: 20px;
+    width: 12px;
+    height: 12px;
     fill: #ffffff;
 }
 
@@ -120,8 +147,6 @@ input {
         right: 20px;
     }
 }
-
-
 .graphic_icon {
     display: flex;
     justify-content: center;
