--- src/pages/SprintView.js
@@ -1,14 +1,0 @@
-import SectionView from '../components/SectionView';
-import Container from '../components/Container';
-
-const SprintView = () => {
-  return (
-    <SectionView>
-      <Container>
-        <p>sprints</p>
-      </Container>
-    </SectionView>
-  );
-};
-
-export default SprintView;
