+++ src/pages/SprintView/index.js
@@ -1,0 +1,1 @@
+export { default } from './SprintView';
